requirements.txt
# ============================================================================
# File: requirements.txt (Root directory)
# ============================================================================
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python-headless==4.8.1.78
numpy==1.24.3
scikit-image==0.21.0
Pillow==10.1.0